Prepare release 0.37: various minor
- Add LICENSE for t/98_kwalitee.t (only specified in EUMM rule)
- Add CONFIGURE_REQUIRES => DBI [cpan #110188]
- Fix little spelling error [cpan #85437]
- Silence void* to int cast warning in func.c
- Improve kwalitee test

+++ func.c
@@ -35,7 +35,7 @@ static void minmaxFunc(sqlite_func *context, int argc, const char **argv){
   int mask;    /* 0 for min() or 0xffffffff for max() */
 
   if( argc==0 ) return;
-  mask = (int)sqlite_user_data(context);
+  mask = (int)(unsigned long)(sqlite_user_data(context));
   zBest = argv[0];
   if( zBest==0 ) return;
   if( argv[1][0]=='n' ){
@@ -523,7 +523,7 @@ static void minmaxStep(sqlite_func *context, int argc, const char **argv){
   }else{
     xCompare = strcmp;
   }
-  mask = (int)sqlite_user_data(context);
+  mask = (int)(unsigned long)(sqlite_user_data(context));
   assert( mask==0 || mask==-1 );
   p = sqlite_aggregate_context(context, sizeof(*p));
   if( p==0 || argc<1 ) return;
